+++ src/cbor.h
@@ -28,3 +28,8 @@
 #endif
 
 #include "php_cbor.h"
+
+#ifndef min
+#define min(a, b)  ((a) < (b) ? (a) : (b))
+#define max(a, b)  ((a) > (b) ? (a) : (b))
+#endif